add union & bitfield to sv_flags for easier dump analysis
Calling Perl_dump_sv from a C debugger on demand on a SV* is not always
possible, if all threads in the process are frozen, it segv'ed or is a
dump file (not a process), or PerlIO/interp struct is in a partial state
and can't print to the console. So create this bitfield on compilers that
support anon unions to make looking at the sv head easier. I expect more
compiler will enable HAS_ANONFIELDS on them in the future.