reduce # of calls to errno
On some OSes like Win32 and Linux, errno is a macro, which contains a
func call that returns a int *. Excessive getting/calling of it is
undesirable. Use stack vars, factor out, reposition the errno to the
branch it is used in, or change the end of a branch to be more friendly
to being merged by CC. In win32.c, neither GCC 02 or VC 01 factored out,
or attempted to merge the func calls when assigning to errno, each time
errno was written, machine code showed a unique call. So do it explictly.

No attempt was made at fixing any errno bugs, such as lack of setting
errno or what constant errno is being set to. This search for errno
callers was not exhaustive. More will need to be done. VC2003 32bits
perl521.dll's .text section dropped from 0xC281F to 0xC278f bytes of
machine code with this patch.