introduce PL_w32_osver
Calling GetVersionEx in an XS module is a common activity according to CPAN
grep. The only RW static global var in Win32:: is an OSVERSIONINFO struct.
Win32:: is often loaded by Win32 Perl users for various reasons. So to
decrease its load time, and remove all perl caused (non-CC caused) RW
static data from Win32::, export core's OSVERSIONINFO struct as a data
export. For XS modules in general, the DLL overhead of the perl core data
export and overhead of using GetVersionEx are the same. 2 ptr sized values
and 1 null terminated string (symbol name). So if a XS module stops
importing GetVersionEx and instead imports PL_w32_osver, the space usage
in the DLL is basically identical. If for that XS module, OSVERSIONINFO was
the only RW global data, then it is a win in memory. Obviously it is a win
in execution time to not call GetVersionEx and do pointer derefs instead.
The DLL loader will have the same work to do resolving GetVersionEx and
PL_w32_osver symbols, but GetVersionEx has to execute later but
PL_w32_osver doesn't. Since this data can never change between reboots,
it is process global and not interp global for ithreads/psuedofork
purposes. An unimplemented idea would be to prefill the OSVERSIONINFOEX
struct with the data from PC that the perl core was built on, and use
.data section symbol ordering (a "map" file) to stick PL_check, PL_ppaddr
and PL_w32_osver in 1 DLL/OS page and have Windows do interprocess COW
on that page. VC doesn't treat the imported data * in the IAT as const
and will reread the IAT entry of the DLL between function calls. To avoid
double indirection over and over (something like
"eax = *(U32*)(eip+__imp__PL_w32_osver);
if(*(U32*)(eax+dwMajorVersion) >= 6) {")
create dW32OSVER and w32_osver as C auto caches of PL_w32_osver.

It has not been decided what to do with Windows 8.1's freezing of the data
GetVersionEx reports. This API is designed to be possibly included in
ppport.h. See how Win32:: chooses to use its per module GetVersionEx or
the global one.