WIP dont VOL protect vars from setjmp on Win32
Needs more research, I dont think this is 100% safe on Win32. Test with GCC
and write a shadowing VOL version with DebugBreak asserts to prove it.
Some of the vars might not be modifying in between setjmp and longjmp or
something.

This decreased perl521.dll VC 2003 -O1 .text section from 0xCE723 to
0xCE6F3 bytes of machine code.