remove all usage of dl_undef_symbols on platforms where its a nop
Only freemint uses it. Change dl_undef_symbols to be a vararg on platforms
that dont implement it, since a vararg skips the computing the number of
incoming items, and the conditional branch testing the item count, and
the call to Perl_croak_xs_usage. This makes the machine code smaller for
dl_undef_symbols on platforms that dont implement it. Retain returning
empty list for back compat. CODE: is more efficient than "SP-= items;"
that PPCODE blocks do, since CODE: will just copy mark to sp, not caring
how many items ahead, sp was, compared to mark.