ExtUtils::Constant stop using SCALAR of *{"_<$filename"} debugging glob
This optimization was added in commit 6736a914b8
"For PROXYSUBS error messages, generate more efficient code for -Uusethreads."

On unthreaded Perl, CopFILE is implemented with debugging path GVs, the
same ones as for storing perl source code. Getting the GvNAME from the
name HEK involves adding the offset from the start of the HEK to the start
of the string part. If you use the SCALAR for the filename, you dont need
to add 8+2 to the pointer, and that saves an cpu instruction. I think that
was NC's point. To allow a future patch that stops creating the
SCALAR part of the debugging glob unless actually using a perl debugger,
CopFILESV can't be used anymore, since in the future it will almost always
be NULL. This patch is related to [perl #125296] and [perl #125512].

Initially undoing NC's change (CopFILESV->CopFILE0 caused
XS_Fcntl_AUTOLOAD's size in machine code on VC2003 -O1 to increase from
0xCB to 0xD1 bytes on unthreaded Win32 perl (the optimization has no effect
with threads). After additional optimizations in AUTOLOAD, the final
size of AUTOLOAD xsub is 0xC0. The cost of reverting the NC optimization
is 6 bytes. Saving just 1 scalar head (no body, no PV) is
16 bytes on 32 bits, so the cost is in the definetly in benefit for
reverting the optimization.

AUTOLOAD
-dont store cop in a C auto around multiple function calls, reduce liveness
-factor out the newSVpvf and 4 args out into 1 block. VC optimizer, which
 isn't the smartest one out there, put 2 identical except for the format
 string branches in machine
-except for arg sv, the other 3 args are calculated in a function call
 free area after hv_exists_ent() executes. Due to C not standardizing the
 order of execution (sequence point) of func args, compute fmt string in
 a separate statement to prevent CopFILE and CopLINE from being stored in
 C autos or non-vols regs while the fmt string is being computed which
 involves func calls

get_missing_hash, previous size 0x76, after size 0x65
-dont deref var ref repeatedly, after each function call the CC must
 reread it
-reorder the call to newHV so the retval of newHV does not need to be
 stored to a non-vol register to be saved around the func call in
 SvUPGRADE, also the retval of newHV will become the retval of
 get_missing_hash without a move cpu instruction