modify behavior of SCALAR in *{"_<$filename"} dbg glob for mem saving
GvFILE's contents is different based on threads and no threads, probably
due to the design of CopFILE in threads and no threads. This difference
affects the ability to optimize things and what is a COW or refcount of
what.

1. Never create the SCALAR part in *{"_<$filename"} debug glob unless
running a perl debugger.

2. When running under perl debugger with threads, use HEK COW from GvFILE
HEK to store filename in SCALAR of *{"_<$filename"} for debugger. Saves
one file path PV (which is a path to a .pm) for each .pm when debugging.

3. GvFILE is very rarely used in core, it is used just once, soley to
provide the "Name "%s::%s" used only once: possible typo" warning.
GvNAME_HEK is a "_<" decorated HEK and was not usable before. Instead
modify GvFILE to be a "_<" decorated HEK instead of a undecorated
HEK on no threads. This allows curcop's GvNAME_HEK to become GvFILE HEKs
with just a ++ on refcount and skipping a hash lookup in newGP, and saves
mem by not creating a new undecorated HEK to store the path.

See mem statistics in future RT ticket for this patch.

See also [perl #125492]. Related to [perl #125296].